study4/requirements.txt
anthropic>=0.50.0
openai>=1.50.0
google-genai>=1.0.0
yfinance>=0.2.50
requests>=2.32.0
pyyaml>=6.0.1
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
rich>=13.0.0
